Time: 58 ms (40.11%) | Memory: 59.3 MB (45.76%) - LeetSync

+++ 2665-minimum-time-to-repair-cars/minimum-time-to-repair-cars.cpp
@@ -5,7 +5,7 @@ private:
     bool canRepair(vector<int>& ranks , ll mid , int cars){
         ll can = 0 ;
         for(auto& it : ranks){
-            can += (sqrt(mid * (1.0) / it)) ;
+            can += (ll)(sqrt(mid * (1.0) / it)) ;
         }
         return can >= cars ;
     }
